don't move the completer further to the left than the left border of the editor
Fixes bug 1688
Note: in the current implementation, the completer (widget) is a child of the
window. Therefore, it will be clipped at the window borders. This may become an
issue for large completer boxes. We should never clip the left of the completer
because it contains the most important information.

Further possible improvements:
Currently we don't go further to the left than the left of the editor. This is
simple and safe. One could go further to the left up to the window border, but
this requires handling at a higher-level code position. Alternatively, one could
make the completer float independent of the window borders. Still, both cases
would need proper shifting of large boxes to guarantee that a maximum of the
completer width is visible while not beeing left-clipped.