Only insert \caption{} once
When using the "Insert Graphic" wizard, if only a short caption is
provided, TeXStudio currently inserts an extra, empty \caption{}.

This happens because inserting a label requires a caption and tries to
add it if the caption is empty. Of course, in the case that only a short
caption is provided, we end up with two \caption{}s.

By checking that both the long and short caption is empty, we make sure
only one caption is inserted.