Don't show link overlay for \begin and \end
The current implementation allows to click on the environment
name in \begin{name} to jump to the corresponding definition.
Clicking on \begin does nothing, so it might be misleading
to highlight it with a link overlay.

Alternatively, one could (in the future) provide a link overlay
for the entire \begin{name} expression.