fix #1768
The issue is that % in latex is ambigious. The qce parser works with
static regex where txs has become more flexible.
The folding mechanism/match highlight still is based on qce, so comments
are not handled properly.
This "fix" is a small work-around a probably often occuring nuisance,
but this should be fixed properly sometimes