Removed Features (Read the description)
Removed the README feature showcase and made the appeal to visit the website more prevalent.
Considering the website **already** has the exact same showcase **and** the 'Find out more at our website' was pointing the user away from the README in the first place, its sensible to choose one location for the showcase.