XKeyboard
Move the macro from the `X11Exception.h` file to `XKeyboard.cpp` as it is the only file that is and probably ever will use it.
While moving it, also changed it's name to `assert` and thus marking it as local -> added `#undef`.
Added error msg for the macro uses that didn't have any msg yet, might have to be reworded in a future rework of the `XKeyboard.cpp` file.