Make reading configuration window size more rebust
Make reading configuration window size more rebust, so it can deal with incorrect x and y values for the top left corner of the window, as are being returned under Wayland. The changes makes sure that even if the position of the window is nonsense, the window size can get restored (as long as the window fits the screen).