Update usermanual_en.html (#2578)
* Update usermanual_en.html

Label #SHORTCUTS refers to section 4.13 Keyboard shortcuts, not 4.12 Synopsis of the TeXstudio command

* Update usermanual_en.html

It seems that the label was TENTATIVE-FEATURES all the time. And while the title of the fifth chapter was _Tentative Features_, the index used _Experimental Features_ for this chapter. I changed every of these _tentative_ words to word _experimental_.