fix handling ai responce where text=""" is used to denote answer
txs sends selected text as "text="""...""" and adds the actual query.
Now mistral answers with text="""..."" and some more remarks.
txs now inserts the content of text="""...""" when "insert result" is
selected.
This is helpful for queries like change letter a to b in text (text
being the selected text)